0.10.3 Fixes multiple error where multiple BUkcaroo widgets rendered improperly in marimo-wasm
attempt at fixing the unstyled DFViewer that only shows up in marimo-WASM for two or more buckaroo widgets.  I think it was because of how theme was defined as a global vs function local variable and the difference between status bar and dfviewer. it was defined inside the StatusBar function and at module level for DFViewerInfinite.  moving both to inside the function and wrapping in useMemo (#380)