Feat/execution framework lazy (#439)
The basis for the new execution framework that allows reliable execution of analytics in a different process. This PR has the main types for the execution framework
This adds the LazyInfinitePolarsBuckarooWidget that works with lazy frames and computes summary stats in the background.
Adds sqlite base executor log and file cache along with tests.